% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact2slim.R
\name{exact2slim}
\alias{exact2slim}
\title{Convert a set of GO exact terms to GO slim terms.}
\usage{
exact2slim(Exact_terms, Slim_collection = NULL, Category = "BP",
  Exclude_self_slim = F, Unique = T)
}
\arguments{
\item{Exact_terms}{vector of query GO exact terms.}

\item{Slim_collection}{the \code{OBOCollection} object, if not provided, it will use the generic GO slim collection downloaded from \url{"http://geneontology.org/ontology/subsets/goslim_generic.obo"}.}

\item{Category}{a character specifying the gene ontology category, can be one in "BP", "CC", and "MF", default "BP".}

\item{Exclude_self_slim}{whether to remove the slim terms of the Category it self, default FALSE.}

\item{Unique}{wheather to unique the mapped GO slim term of each GO exact term.}
}
\value{
a \code{data.frame} indicating the mapping index between query GO exact terms and its GO slim terms.
}
