% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2goslim.R
\name{gene2goslim}
\alias{gene2goslim}
\title{A function that map Gene IDs to GO slim terms.}
\usage{
gene2goslim(Gene_ID, Gene_key_type = "ENTREZID", OrgDB, Category = "BP",
  Drop_zero_match = F, Exclude_self_slim = F)
}
\arguments{
\item{Gene_ID}{gene ids to query.}

\item{Gene_key_type}{a character specifying the type of the ID, the available types of the keys can be find using \code{keytypes(org.Hs.eg.db)}, default "ENTREZID".}

\item{OrgDB}{an \code{OrgDb} object defined by AnnotationDbi.}

\item{Category}{a character specifying the gene ontology category, can be one in "BP", "CC", and "MF", default "BP".}

\item{Drop_zero_match}{whether to drop the genes that mapped to no terms at all, default FALSE.}

\item{Exclude_self_slim}{whether the GO slim terms of its own category i.e. remove terms of c("GO:0008150","GO:0005575","GO:0003674").}
}
\value{
a \code{data.frame} that define the mapping relationship.
}
