% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOslimEA.R
\name{GOslimEA}
\alias{GOslimEA}
\title{Vectorized function to enrich GO slim terms with a list of gene sets against a common background.}
\usage{
GOslimEA(gene_set, back_ground, orgDb, category = "BP",
  gene_key = "ENTREZID", min_gs_count = 10, max_gs_count = 500,
  exclude_root_slim = T, exclude_zero_slim = T, EASE_Score = F,
  pvalue_correction = "BH", interpret_term = F, show_gene_name = F)
}
\arguments{
\item{gene_set}{a list of character vectors contains gene IDs of the query gene sets.}

\item{back_ground}{a character vector contains the IDs of the background genes.}

\item{orgDb}{an \code{OrgDb} object defined by AnnotationDbi.}

\item{category}{a character specifying the gene ontology category, can be one in "BP", "CC", and "MF", default "BP".}

\item{gene_key}{a character specifying the type of the gene ID, the available types of the keys can be find using \code{keytypes(org.Hs.eg.db)}, default "ENTREZID".}

\item{min_gs_count}{GO slim term minimum number of occurence in the query gene set; default 10.}

\item{max_gs_count}{GO slim term maximum number of occurence in the query gene set; default 500.}

\item{exclude_root_slim}{whether to drop the GO slim terms that are roots of "BP", "CC", or "MF" at the very first, default TRUE.}

\item{exclude_zero_slim}{whether to exclude the involvement of the genes that belong to no GO slim terms at all, default TRUE (increase power).}

\item{EASE_Score}{whether or not use EASE score method. (a more conservative hypergeomatric test calculation used in DAVID)
for more details please refer to \url{https://david.ncifcrf.gov/helps/functional_annotation.html#fisher}, default FALSE.}

\item{pvalue_correction}{method used for multiple hypothesis adjustment, can be one in "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", and "none".}

\item{interpret_term}{whether to let the GO term readable, default FALSE.}

\item{show_gene_name}{whether to attach readable gene names for each GO term, default FALSE.}
}
\value{
a \code{list} of \code{data.frame} for GO slim enrichment result.
}
\details{
\code{GOslimEA} conduct GO slim EA analysis using Fisher's exact test / permutation test on GO slim terms. GO slim terms refer to the reduced GO terms that are roots in GO hierachical structures.
}
