% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea}
\alias{gsea}
\title{permutation test based gene set enrichment analysis given count of each terms and total gene numbers for the gene set and background.}
\usage{
gsea(freq_gs, freq_bg, gs_total_gene, bg_total_gene, adj_method, ease)
}
\arguments{
\item{freq_gs}{The named vector of term frequencies in gene set.}

\item{freq_bg}{The named vector of term frequencies in background.}

\item{gs_total_gene}{Total gene number in gene set.}

\item{bg_total_gene}{Total gene number in back ground.}

\item{adj_method}{the p adjust method.}

\item{ease}{To calculate EASE score or not.}
}
