% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2go.R
\name{gene2go}
\alias{gene2go}
\title{A function that map Gene IDs to GO terms.}
\usage{
gene2go(Gene_ID, Gene_key_type = "ENTREZID", OrgDB, Category = "BP",
  Drop_zero_match = F)
}
\arguments{
\item{Gene_ID}{gene ids to query.}

\item{Gene_key_type}{a character specifying the type of the ID, the available types of the keys can be find using \code{keytypes(org.Hs.eg.db)}, default "ENTREZID".}

\item{OrgDB}{an \code{OrgDb} object defined by AnnotationDbi.}

\item{Category}{a character specifying the gene ontology category, can be one in "BP", "CC", and "MF", default "BP".}

\item{Drop_zero_match}{whether to drop the genes that mapped to no terms at all, default FALSE.}
}
\value{
a \code{data.frame} that define the mapping relationship.
}
